initializing registry users in the DB only once for all ITs

+++ registry-integration-tests/src/test/java/org/gbif/registry/database/TestCaseDatabaseInitializer.java
@@ -138,9 +138,11 @@ public class TestCaseDatabaseInitializer implements BeforeEachCallback {
         }
       }
 
-      connection
-          .prepareStatement("TRUNCATE " + String.join(",", tablesNotChallengeCode) + " CASCADE")
-          .execute();
+      if (!tablesNotChallengeCode.isEmpty()) {
+        connection
+            .prepareStatement("TRUNCATE " + String.join(",", tablesNotChallengeCode) + " CASCADE")
+            .execute();
+      }
 
       connection.commit();
